+++ lib/aoc_cli.rb
@@ -1,5 +1,4 @@
 require 'colorize'
-require 'json'
 require 'sqlite3'
 
 require_relative 'aoc_cli/errors'
@@ -17,6 +16,7 @@ require_relative 'aoc_cli/tables'
 
 module AocCli
 	Metafile = AocCli::Files::Metafile
+	Prefs    = AocCli::Files::Config::Prefs
 	Validate = AocCli::Interface::Validate
 	E = AocCli::Errors
 end