+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-17-02-21
-	0.1.2
+## 17-02-21
+### 0.1.2
 	- Add changelog
 	- Add user-agent
+
+## 20-02-21 
+###1.0.0
+	- Add calendar tables
+		- Star progress now based off this database rather than meta json
+	- Add cat of calendar from any directory
+	- Auto generate config file
+	- Better key validation
+	- Better default alias handling
+		- List all aliases
+	- Rehaul of year initialisation
+	- Expand config integration
+		- Ability to turn on/off calendar writing
+		- Turn on/off leaderboard stats in calendar file
+	- Git integration
+	- General refactor